Added date/time(incl. conversions), location/distance on horizontal split text views

+++ app/src/main/java/com/example/android/quakereport/Earthquake.java
@@ -1,19 +1,26 @@
 package com.example.android.quakereport;
 
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
 /**
  * Created by nagma on 07.28.16.
  */
 public class Earthquake {
     private String mLocation;
     private double mMagnitude;
-    //private String mDate;
-    private int mTime;
+//    private String mDate;
+    private long mTime;
 
-    public Earthquake(double magnitude, String location, int time){
+    public Earthquake(double magnitude, String location, long time){
         this.mLocation = location;
         this.mMagnitude = magnitude;
         this.mTime = time;
     }
+
+//    public Earthquake(String date){
+//        this.mDate = date;
+//    }
 
     public String getLocation(){
         return mLocation;
@@ -23,7 +30,7 @@ public class Earthquake {
         return mMagnitude;
     }
 
-    public int getTime(){
-        return mTime;
-    }
+    public long getTime(){ return mTime; }
+
+//    public String getDate(){ return mDate; }
 }
